Update createIndexes example for PHP
This example was linking to the current PHP driver but showing syntax from the legacy driver. Updated to correctly link to the current PHP library API. Keeping "PHP Driver" language in place for consistency, despite the fact that this is handled by the library.

These examples may change later with DRIVERS-448, but this is a quick fix.